Separate the collections edit groups header related components #37
